+++ LibraryManagement.Application/ViewModels/BookViewModel.cs
@@ -5,9 +5,9 @@
         public string Title { get; private init; }
         public string Author { get; private init; }
         public string Isbn { get; private init; }
-        public int PublicationYear { get; private init; }
+        public ushort PublicationYear { get; private init; }
 
-        public BookViewModel(string title, string author, string isbn, int publicationYear)
+        public BookViewModel(string title, string author, string isbn, ushort publicationYear)
         {
             Title = title;
             Author = author;
